# lib/api-zod/src/generated/types/atlasAlphaScore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Atlas Alpha - Quant Trading Signal Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AtlasAlphaScoreDirection } from './atlasAlphaScoreDirection';
import type { AtlasAlphaScoreLabel } from './atlasAlphaScoreLabel';
import type { AtlasAlphaScoreTimeHorizon } from './atlasAlphaScoreTimeHorizon';

export interface AtlasAlphaScore {
  overall: number;
  label: AtlasAlphaScoreLabel;
  trendScore: number;
  momentumScore: number;
  volumeScore: number;
  optionsScore: number;
  relativeStrengthScore: number;
  marketRegimeScore: number;
  bullishProbability: number;
  bearishProbability: number;
  confidenceScore: number;
  riskScore: number;
  direction: AtlasAlphaScoreDirection;
  timeHorizon: AtlasAlphaScoreTimeHorizon;
  expectedMovePercent: number;
  indicatorsAgreeing: number;
  totalIndicators: number;
  signalNarrative: string;
}